+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/utils/TypeUtils.java
@@ -24,6 +24,7 @@ import io.ballerina.runtime.internal.TypeChecker;
 import io.ballerina.runtime.internal.types.BArrayType;
 import io.ballerina.runtime.internal.types.BFiniteType;
 import io.ballerina.runtime.internal.types.BType;
+import io.ballerina.runtime.internal.types.BTypeReferenceType;
 
 import static io.ballerina.runtime.api.PredefinedTypes.TYPE_ANY;
 import static io.ballerina.runtime.api.PredefinedTypes.TYPE_ANYDATA;
@@ -134,4 +135,13 @@ public class TypeUtils {
     public static boolean isSameType(Type sourceType, Type targetType) {
         return TypeChecker.isSameType(sourceType, targetType);
     }
+
+    public static Type getReferredType(Type type) {
+        Type constraint = type;
+        if (type.getTag() == TypeTags.TYPE_REFERENCED_TYPE_TAG) {
+            constraint = getReferredType(((BTypeReferenceType) type).getReferredType());
+        }
+        return constraint;
+    }
+
 }
